requirements.txt
asgiref==3.8.1
Django==4.2.22
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.14.0
